+++ ve_brain/ve_brain/__init__.py
@@ -22,16 +22,17 @@ from .contracts import (
 from .strategy_contract import ValidationStatus, can_reach_n6, can_execute_real, TradeProposal
 from .regime_routing import (
     SemanticRegime, RawAxes, applicable_regimes, regime_fingerprint, StrategyContract, StrategyRegistry,
-    StrategyRouter, RoutingMode, EligibilityDecision, n4_triggers_breakout, requires_true_range,
-    strategy_policy_fingerprint, DuplicateStrategyPolicyError,
+    SealedRegistry, catalog_hash, StrategyRouter, RoutingMode, EligibilityDecision, n4_triggers_breakout,
+    requires_true_range, strategy_policy_fingerprint, DuplicateStrategyPolicyError,
+)
+from ._canonical_catalog import (
+    CANONICAL_CATALOG_VERSION, CANONICAL_CATALOG_HASH, CANONICAL_STRATEGIES,
 )
 from .ev_engine import run_ev, ENGINE_VERSION, EV_ENGINE_SOURCE_COMMIT
 from .fingerprint import (
     decision_fingerprint, data_identity, require_comparable, compare_decisions, NonComparableDecisionError,
 )
-from .n6 import (
-    decide_n6, register_canonical_strategy, reset_canonical_registry, set_registry_available,
-)
+from .n6 import decide_n6
 
 __version__ = VE_BRAIN_VERSION
 
@@ -41,8 +42,8 @@ __all__ = [
     "build_info", "ReasonCode", "DecisionRequest", "DecisionResponse", "ProbabilityInputs", "OutcomeCell",
     "HierarchyLevel", "validate_request", "validate_response", "SchemaValidationError", "INPUT_CONTRACT_ID",
     "OUTPUT_CONTRACT_ID", "ValidationStatus", "can_reach_n6", "can_execute_real", "TradeProposal",
-    "SemanticRegime", "RawAxes", "applicable_regimes", "regime_fingerprint", "data_identity", "N1_CONTRACT_VERSION", "ROUTER_VERSION", "RAW_AXIS_SCHEMA_VERSION", "ELIGIBILITY_POLICY_VERSION", "StrategyContract", "RANGE_STRATEGY_ROUTING", "RANGE_BLOCKER", "StrategyRegistry", "StrategyRouter", "RoutingMode",
+    "SemanticRegime", "RawAxes", "applicable_regimes", "regime_fingerprint", "data_identity", "N1_CONTRACT_VERSION", "ROUTER_VERSION", "RAW_AXIS_SCHEMA_VERSION", "ELIGIBILITY_POLICY_VERSION", "StrategyContract", "RANGE_STRATEGY_ROUTING", "RANGE_BLOCKER", "StrategyRegistry", "SealedRegistry", "catalog_hash", "StrategyRouter", "RoutingMode",
     "EligibilityDecision", "n4_triggers_breakout", "run_ev", "ENGINE_VERSION", "EV_ENGINE_SOURCE_COMMIT",
     "requires_true_range", "strategy_policy_fingerprint", "DuplicateStrategyPolicyError", "decision_fingerprint", "require_comparable", "compare_decisions", "NonComparableDecisionError", "decide_n6",
-    "register_canonical_strategy", "reset_canonical_registry", "set_registry_available",
+    "CANONICAL_CATALOG_VERSION", "CANONICAL_CATALOG_HASH", "CANONICAL_STRATEGIES",
 ]